SMAR-19
Adding a new field to a normal user is admin
Boxes for entering email and password.
Data search in the DB of the entered email and password.
When you click the button to connect, the entered data is checked to see if it is validated against the data retrieved from the top priority DB. And if an admin field is equal to true